api/requirements.txt
supabase==1.0.3
requests==2.31.0
python-dotenv==1.0.0
# 如果你要用到其他库，也需要添加在这里